Avoid RecordNotFound error on destroy

+++ plugins/orders/controllers/myprofile/orders_plugin_item_controller.rb
@@ -35,7 +35,8 @@ class OrdersPluginItemController < MyProfileController
   end
 
   def destroy
-    item = OrdersPlugin::Item.find params[:id]
+    item = OrdersPlugin::Item.where(id: params[:id]).first
+    return unless item
     @offered_product = item.offered_product
     item.destroy
   end
